Refactor `SigArgs` handling into separate files.
Moved `SigArgs` struct and related parsing logic from `stream.cpp` to new dedicated files `signature_arguments.cpp` and `signature_arguments.h`. This improves modularity, making the codebase easier to maintain and extend. Updated the CMake configuration accordingly.
